app/requirements.txt
kafka-python
cassandra-driver
pandas
fastapi[standard]